pkg/dt: Move kernel-independent parts of parsing to fdt.go
